+++ src/main/java/com/mju/lighthouseai/domain/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.mju.lighthouseai.sample.entity;
+package com.mju.lighthouseai.domain.user.entity;
 
 import com.mju.lighthouseai.global.entity.BaseEntity;
 import jakarta.persistence.Column;
@@ -34,6 +34,12 @@ public class User extends BaseEntity {
     private String nickname;
 
     @Column(nullable = false)
+    private String birth;
+
+    @Column(nullable = false)
+    private String profile_img_url;
+
+    @Column(nullable = false)
     @Enumerated(EnumType.STRING)
     private UserRole role;
 
@@ -42,12 +48,16 @@ public class User extends BaseEntity {
         final String email,
         final String password,
         final String nickname,
-        final UserRole role
+        final UserRole role,
+        final String birth,
+        final String profile_img_url
     ) {
         this.email = email;
         this.password = password;
         this.nickname = nickname;
         this.role = role;
+        this.birth = birth;
+        this.profile_img_url = profile_img_url;
     }
 
     // 롤 변경 메소드
