+++ src/main/java/com/mju/lighthouseai/sample/service/impl/TestServiceImpl.java
@@ -3,12 +3,12 @@ package com.mju.lighthouseai.sample.service.impl;
 import com.mju.lighthouseai.sample.dto.service.TestCreateServiceRequestDto;
 import com.mju.lighthouseai.sample.dto.service.TestReadResponseDto;
 import com.mju.lighthouseai.sample.entity.Test;
-import com.mju.lighthouseai.sample.entity.User;
 import com.mju.lighthouseai.sample.exception.NotFoundTestException;
 import com.mju.lighthouseai.sample.exception.TestErrorCode;
 import com.mju.lighthouseai.sample.mapper.entity.TestEntityMapper;
 import com.mju.lighthouseai.sample.repository.TestRepository;
 import com.mju.lighthouseai.sample.service.TestService;
+import com.mju.lighthouseai.domain.user.entity.User;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
